Add LINK instruction to workaround hardware bug in Link register.
The Link register is wired incorrectly in hardware so that during JALR, it
sometimes captures the value of PC after the jump executes. (It's racey.)
Add the LINK instruction so that we can break JALR into two opcodes: LINK
and JMP. This allows us to capture the value of PC before performing the
jump.

Note that it's now necessary to increment the return address before the
JMP to return from a function.