Add "eval" statement to the Snap language.
The "eval" statement is followed by an expression. This expression is
compiled to code which evaluates the expression at runtime, placing the
result into the A register.

The idea here is to add a statement which allows expression compilation to
occur, and without bringing in any other features. For example, tying this
to the "let" statement meant I had to start thinking about allocating
variables at runtime. Tying this to the "return" statement meant I had to
start thinking about returning from a function call, the stack, &c.