Restrict the source and destination registers allowable for BLT.
The BLT instruction is intended to accelerate the performance of block memory copies between data RAM and device memory or instruction RAM. It asserts both the INUV and INXY control signals and also performs a bus transfer between a specified source and destination register. Restricting to only the P and M pseudo-registers would greatly reduce the number of required opcodes without much decreasing the utility of the instruction.