Fixed and dynamic arrays now have a "count" property for the length.
Added "Get" properties which can get a property on a value. Currently, the
only implemented property is the "count" property for fixed arrays and
dynamic arrays.

The compiler instrinsic function "length" has been removed.

The standard library (such as it is) now uses the string "count" instead of
calling the "length" function.