Fix unit tests for getting array count.
Fix the following tests:
* testGetLengthOfLiteralArray
* testGetLengthOfArrayThroughIdentifier
* testGetLengthOfDynamicArray

When getting the count of an array, first evaluate the array expression
and allow the result to be written to a temporary. (This is necesary in
case evaluation has side effects.) Then, for fixed-size arrays, emit code
to copy the array count to a new temporary. For dynamic arrats, emit code
to extract the array count and copy that to a new temporary.