Special handling of self parameter in struct member function calls.
Given a struct member function call on a struct instance, check to see if
the first parameter of the function is a pointer or const-pointer to the
struct type in question with the name "self". If so then that parameter
should be populated with a pointer to the struct instance object instead
of coming out of the call parameter list as usually happens.