Add the beginning of a module system. Use it to import stdlib.
Currently, the first time a module is imported, the compiler will load the
module source code from file, parse it, and cache resulting AST. This AST
is injected into the program at the site of the import statement.

This is super crude, and only a little bit better than textual inclusion.
However, it allows me to start iterating on the module system and syntax.