Fix bug which broke code which tried to get an array from a struct and subscript it.
Subscripting an array used to assume the subscriptable expression was a
simple identifier. It would downcast to an Identifier and then try to look
up the array base address from the symbol table. However, when the
subscriptable is a Get expression, this is not possible.

Now, the subscriptable expression is compiled and evaluated in an lvalue
context to get the array base address. (But only if the Get expression
evaluates to an array type. This change doesn't tackle the other cases.)