Added `bitcastAs' operator to perform unsafe bitcasts such as changing the type of a pointer.
This allows an implementation of something like an interface by casting a
pointer to the concrete struct to a pointer to a vtable struct. So long as
the vtable is the first element of the concrete struct, the field layout
in memory is such that making member function calls through the casted
pointer will work as expected.

Also, the instance of the concrete struct needs to have a vtable properly
setup ahead of time. This requires the use of `bitcastAs' too since the
types of the self parameter are different.