Fix bug where range in subcript causes unsupported expression error.
This involves rolling back changes in 5b03f71d26d9928d5dfe6f127eb8fc11bfb29513.
As a consequence, it is impossible again to return a struct that is too
large to fit into a compiler temporary.

The bug was that the byte copy that was being used to avoid the creation
of that compiler temporary needed to get the lvalue of the expression on
the right-hand side. However, that right-hand expression has no lvalue
when it is a StructInitializer as is the case for a Range.