The body of the COPY IR instruction is extracted to new procedure.
The COPY IR instruction can generate a call to a new procedure to execute
the byte copy loop. The return is done via the link register, to avoid
involving the stack.

If number of bytes to copy is less than spme threshold (6 bytes) then it is
more efficient, in terms of code size, to have the compiler unroll the
loop. Else, generate code to call that procedure.