Begin work on an optimizer for Crackle IR code.
The compiler generates some really unfortunate code. It can be greatly
improved by rewriting instructions based on what can be statically
determined about the contents of memory. For example, we know the value of
memory when a STORE-IMMEDIATE is issued. We can then use this information
to determine the source and destination addresses in indirect copies and
replace them with direct copy equivalents.