Optimize compiled programs by propagating constants in assembly code.
Immediate values which are loaded into registers and memory can be tracked
across a basic block. We can skip instructions which would have no effect
such as LI X, Something when we already know that X contains Something.
Similar tracking can be done for the contents of RAM.

Instructions can be rewritten from MOV to LI when we can statically
determine the value that will be assigned. The benefits of this change will
be apparent once dead store elimination has been implemented.