The /J signal should be presented to PC when the ALUResult is latched.
If we latch the control signal and latch the ALUResult in EX/MEM and then
present the register outputs to the PC stage then the signals will be
stable for an entire clock cycle.

If we take then at the exit point of the EX stage then they'll only be
stable for a few nanosecond before the next clock edge occurs and the ALU
outputs become unstable. Also, let's avoid presenting the /J signal to PC
tens of nanoseconds before the new address is ready.