The right operand can form an immediate value from several bit fields.
This solves the problem of the Store instruction not being able to compute
Ra+Imm as can be done in the Load instruction. Now, it can get those
immediate bits from the unused 'c' field.

Also, Jump isntructions which do not use the 'c' field to specify a link
register can use those bits to form an eight-bit immediate value.