The instruction decoder no longer depends on the /J signal.
Instead, the /J signal feeds into the Hazard Control logic. This can
assert signals to clear the control word register to zero immediately.

Also, the Ctl_ID signals are now inverted wrt to other stages so that
a zero control word means a NOP. Conveniently, we can reset the 74AHCT273
to a zero to effect this.