Compile and simulate HDL for the ATF22V10 wihch implement hazard control.
Compile the PLD source files using galasm for the two ATF22V10 which
implement hazard control. Load the resulting JEDEC files and simulate the
operation of the GALs when we simulate the ID stage of the pipeline.

TODO: There's a bug in combinatorial feedback in the GAL simulator and we
currently fake the logic to calculate `isFlagsHazard'.