Fix combinatorial feedback so we can simulate the STALL and FLUSH equations
The API now allows us to specify a 0, 1, or nil value as input for each
pin, where the value represents a logic value to which the pin is being
actively driven. Pins intended to be output pins are not to be actively
driven by an external device and should be nil.