Add two new control signals to indicate the A and B operands are unused.
Several instructions do not use both the A and B bitfields to specify
operands to be fetched in the ID stage. They might instead use those bits
for an immediate value, for example. For these instructions, the new
control signals are used to give the hazard control unit a better
understanding of instruction semantics so it does not stall.