Add SnapAbstractSyntaxTreeCompiler
The path I've been taking so far has been to implement parts of the
compiler as discrete passes over the AST. Unfortunately, this doesn't
work well with some transformations such as those needed for ForIn and
Match as they both rely on symbols and generate symbols. So, a discrete
symbol generation step would never be the right thing.

Instead, add a new class SnapAbstractSyntaxTreeCompiler which allows
transformations to be run over portions of the tree as it iterates nodes in
one passclas SnapAbstractSyntaxTreeCompiler which allows
transformations to be run over portions of the tree as it iterates nodes in
one pass.