AST nodes add a source anchor to children that lack one
If a child node has a nil source anchor then the parent replaces it with
a copy that includes the parent's source anchor. This makes it possible to
remove a large amount of boiler plate code from the guts of the compiler.
As a trade off, boiler plate is needed for a slew of new withSourceAnchor()
methods. However, these are much easier to ignore.