Avoid unecessary copy when evaluating function call arguments.
If the argument evaluates to a primitive type then we can store the result
directly to the argument pack without first copying to a compiler temp in
memory.

If the argument evaluates to a non-primitive type then evaluating it will
introduce a compiler temporary to hold the result. It's not necessary to
introduce our own temporary here. The expression pushes a register to the
stack which holds the lvalue, and this can be passed along to MEMCPY.