Implement Struct Member Function Call for Tack compiler.
This also involves a change to the way the Struct Member Function Call
is implemented in the entire language. The compiler tries to compile the
function call normally. If that fails on a compiler error then it tries
to rewrite as a Struct Member Function Call and tries again.

The rewrite involves ascertaining the Self object involved in the call
and attempting to automatically convert it to the type of the first
function parameter.