Add RegisterSpiller to rewrite programs to insert spill code
Given a program and a list of live intervals specifying the register
allocation, rewrite the program to insert spill code. This involves
inserting instructions to load a value from memory before each read of a
spilled register, and to store a value back to memory after each write.

If the client has not specified enough reserved temporary registers then
Register Spilling cannot be completed successfully and this returns an
error.