Add RegisterAllocationDriver to perform register allocation
The RegisterAllocationDriver class performs register allocation,
rewriting a program to map an unlimited set of virtual registers to
a limited set of physical registers. When not enough physical registers
are available then the program is rewritten to spill one or more live
intervals to memory.

The driver attempts allocation with zero reserved spill registers,
initially. If this fails then it retries, adding one reserved register,
and so on until the allocation is able to succeed. This ensures that
registers are not wasted by reserving them needlessly.

RegisterAllocationDriver integrates various existing classes such as
RegisterLiveIntervalCalculator, RegisterSpiller, and
LinearScanRegisterAllocator.