Fix bug in Register File where SEML and SEMR were pulled low
The SEML and SEMR inputs of the two dual port SRAMs in the register file
were both being pulled low. This causes reads and writes to always be
directed to the eight semaphores instead of the registers. These semaphores
are not memory and this leads to incorrect operation of the CPU.