Consistent order of parameters to STORE
Up till now we've actually been inconsistent with the order of the
parameters to the STORE instruction. Standardize on a MIPS-style order
where the source register is first and the second and third parameters
form the memory address. The LOAD instruction already follows this
convention.

Also, add a MemoryAddress type so the computer's memory access closures
cannot confuse the value and the memory address at any point in time.