Fix bug with type cast of a literal array in an rvalue expression
The existing machinery for performing a type cast was unable to accept
a literal array because it has no lvalue. Fix this by having the cast
instead rewrite the AST to replace the literal array with one of the
target type. The AST for the individual array elements are modified too.

Add test_EndToEndIntegration_CastArrayOfU16ToArrayOfU8