Fix bug where register allocator fails to find the leading ENTER instr.
This is a bug with the integration of the register allocator. The register
allocator expects the first instruction of a subroutine to be an ENTER
instruction. However, the compiler puts a label declaration first. Fix this
by changing the Subroutine node so it carries it's own name with it. Insert
the label declaration later when the Subroutine node is erased.