Update ProgrammableLogicDecoder to account for new condition code N
BUGS: This change necessitates a corresponding hardware change which has
not been made yet. We still need to route the N condition code through the
CPU. We now also need to update the schematics to account for changes in
the pin out of the ATF22V10s which implement instruction decoding.