Add spot checks for the condition codes generated by signed < and >.
Check a small randomly selected subset of the configuration space to ensure
that the comparisons always generate the expected condition codes. This
tests that the logic for evaluating condition codes is actually correct.
Note that this logic is implemented in DecoderGenerator and in
ProgrammableLogicDecoder. This just tests that the design is sound.