Fix a bug where variables are allocated overlapping offsets in for-in
When compiling a for-in loop, variables declared in the body of the loop
would have stack offsets which overlap the loop variables declared in
the enclosing scope. When the inner variables are initialized, the loop
variables are overwritten and loop iteration breaks. This occurs because
the storage pointer of the inner Block is not updated when the compiler
processes the VarDeclaration nodes in the enclosing Block.

The solution is to have ImplPass reset storagePointer to the expected
value when it enters a block, instead of assuming it's set elsewhere.
This is correct because ImplPass should be the only place where variable
declarations are made, and it traverses nodes in an appropriate order to
avoid issues.

BUGS: I'm concerned that because rvalue() can create compiler temporaries
when compiling expressions in a pass after ImplPass. If this were to happen
then the storagePointer of the inner scopes also becomes out of sync with
the outer scope and it becomes possible for variables to be assigned to
overlapping offsets again.