Fix bug where lifetime of string was too short, leading to crash
The string lifetime ends at the end of the expression for the function
parameter. It's a problem to try to capture a slice to the string and
use it later. Workaround this issue by creating a real stack variable to
hold the string for the duration of the test.

In the future, it would be possible for the compiler to intern string
literals, store them in static storage, and de-duplicate string literals
appearing in the program. This would also solve the lifetime problem by
giving immutable string literals a lifetime to last the entire program.