Function instantiation is deferred until near the end of compilation
This is ground work before function generics can be implemented. The change
permits the compilation of function bodies to be decoupled from the order
the compiler visits the FunctionDeclaration AST node. With this change, the
work is now deferred until after the main body of the program has been
compiled. In the future, the type checker will be able to request
compilation of concrete instances of generic functions it comes across.
(monomorphization)