Add a generic function type which can be examined by the type checker
Add GenericFunctionType which represents a type function, and a
corresponding new case in SymbolType. Also add a new expression,
GenericTypeApplication. This allows type arguments to be applied to a
generic function to yields a concrete function type.

Concrete instantiation of a generic function is not yet implemented so
these changes only affect the type checker.

The parser is unaware of generic types too right now so these features
are not exposed to the programmer.