SnapSubcompilerImplFor uses the type checker to evaluate type expr
It no longer attempts to look up types explicitly by name in the symbol
table and instead uses the type checker to evaluate the identifier as a
type expression. The idea is that this will allow for the use of any
sort of type expression instead of only an identifier.