Implement ImplFor for generic structs
Compiling an ImplFor no longer directly generates code. Instead, types
are added to the environment, functions are enqueued for compilation
later, and sequences of code are enqueued for later processing for
setting up vtables at the beginning of program execution.

BUG: If the ImplFor declaration appears inside a function then the
symbols are added to the function's symbol table. The code sequence for
setting up the vtable will fail because it is inserted outside of the
function's scope.