Remove preamble and replace with SymbolTable.scopePrologue
The preamble in the GlobalEnvironment is not enough. We need to be able to
attach code to any scope for setting up vtables and such from the type
checker. So, remove GlobalEnvironment.preamble and add a new Seq for
SymbolTable.scopePrologue. This prologue is compiled when the Block is
reduced to a Seq. This allows the type checker to continue working within
a paradigm where it only modifies the environment, and cannot generate new
code in the AST.