Add SnapCompilerFrontEndTests and associated machinery
Add new SnapCompilerFrontEndTests test suite. This is an integration
test suite which tests the compiler from Snap textual language down to
the Tack intermediate representation. This IR program is interpreted in
a virtual machine and the state of the virtual machine is inspected to
determine whether the program executed as expected.

This change introduces associated machinery necessary for these tests
such as TackDebugger, which allows introspection of a running virtual
machine.