Fix bug when inferred function return type is void
Fix a bug where we try to assign a `void' type to a variable with inferred
type `void'. The two types are identical and so the assignment was deemed
to be acceptable. However, we later underflow the compiler's register stack
when we try to pop the stack to get the register containing its value.
This leads to a crash in peekRegister(). The solution is to add a case to
ban conversions from `void' to `void'.