Add MEMModuleTestFixture board for testing the standalone MEM module
The standalone MEM module board can be inserted into this test fixture,
where an Arduino can use shift registers to drive the inputs and read the
outputs. This permits the Arduino to run a suite of (WIP) integration tests
on the MEM module. The goal is to be able to fully exercise and validate
its operation independently of the rest of the computer.