MEMModule now has ROM on Bank 0 again
The idea of having a system with no built-in ROM is turning out to be more
complicated than I expected. The boot device is additional complexity that
I'm now reluctant to take on. It's still, after months, not fully designed
and prototyped. Let's eliminate the need for this entire module by having
the computer run from ROM on boot.

The benefits of a ROM-less system are more difficult to achieve than
expected. The device would need a fast sixteen bit ALU chip as in the
program counter. Using simple counters like the 74F161 would bottleneck the
counting sopeed at 12.5MHz and the boot device provides no effective speed
benefit.

Further, the entire scheme is a performance optimization for a machine that
hasn't even been fully constructed yet. I can take on optimization work
later.

So this change sticks a pair of fast ROMS into MEM on Bank zero. A program
can set the Bank register to something else in order to map the built-in
RAM to instruction memory instead.

Also, update MEMModule for KiCad 7.