Revert "Replace call to deprecated unarchiveTopLevelObjectWithData method"
This reverts commit b8c82b066353e0f43a20b7fca58256ad2ff89e3d.

The problem with the previous change was that switching to the recommended
API introduced unit test failures. Switch back to the deprecated
unarchiveTopLevelObjectWithData method until I have a opportunity to
address this issue. It might be time to switch to an idiomatic Codable
serialization scheme, for example.

Also, make changes to Turtle16Computer.restore(from:) to add better error
messages on failure.