MEMModule: ROM ought to be controlled by RAMEN_R signal
The right side of the RAM is connected to the instruction fetch unit.
The left side of the RAM is connected to the system bus.
The ROM should only be enabled in the instruction fetch unit when the right
side of RAM is not selected. Tying it to RAMEN_L is a bug.