ClockModule: apply second bodge for Phi1_1 bug
Must delay the clock signal before it goes into U7A, else a clock
glitch occurs when the Resume button is pressed. If we replace the
input of U7A with the Phi2 clock signal then the glitch goes away.
Phi2 is already buffered with inverters which would provide such
a delay anyway.

Phi1 and Phi2 are slighly out of phase as a result, but the difference
seems like it would be acceptable.