ControlModule: Reorder Ctl_Mid to match the InstructionDecoder{1,2}.pld
The Instruction Decoder PLDs have defined a different pinout for control
signals than expected in the instruction decoder schematics.

The different output pins of the PLD have a different number of available
product terms and so it is not possible to reorder the outputs in the PLD
files. The ordering on the PCB must change to match.