RegisterFileModule: Fix write-back timing bug
Fix the bug where the /WR signal to the SRAMs was arriving at an invalid
time. The signal arrival is now delayed by a series of logic gates. The
signal is deasserted before other inputs change by gating it against the
Phi2 clock.

Note that the main board schematic and PCB has not yet been updated.