ProcessorBoard and MEMModule are updated to put 7028 SRAM into /S mode
This will avoid issues with writes being inhibited by the onboard write
arbitration logic. The programmer will be advised to avoid writing to
memory as it is executing. However, this ought to work now per the data
sheet which states that a simultaneous write and read to the same cell will
work so long as certain timing constraints are obeyed, and we ought to be
doing that now.

This is the same change which was necessary to fix the register file.

NOTE: The MEMModule and ProcessorBoard still have divergent MEM units. The
ProcessorBoard removes the open-drain RDY signal and accepts clock inputs
through the bus connector. The MEMModule PCB does not make this change.