MainBoard is smaller than the ProcessorBoard now
The MainBoard is smaller than the ProcessorBoard now. Additional mounting
holes have been added to allow for smaller boards to be stacked in the
computer.

Additionally, the processor now include pads for parallel termination of
the clock signals coming in from the bus, should this turn out to be
necessary.