Remove CompilerIntrinsicSymbolBinder
Compiler intrinsic types and functions used to be declared by the compiler
and inserted into globalSymbols. The implementation was likewise inserted
by the compiler at a later time. Now, these symbols are defined in the
platform-specific runtime modules. (runtime_Turtle16 and runtime_TackVM)
This module is always implicitly imported into the program.

Changes have been made to TackVirtualMachine to give it special  handling
for inline assembly where the string is "SYSCALL". In this case, run a
subroutine which implements the syscall. In the future, this could be
replaced with full support for inline Tack assembly code.