SymbolTable.storagePointer and highwaterMark replaced by Frame class
The SymbolTable class no longer tracks storagePointer or highwaterMark.
Instead, we introduce a new Frame class which tracks the storagePointer.
Each symbol table has a stackFrameLookupMode which may be set to
either .inherit or .set. In the former case, we retrieve the frame by
looking at the parent symbol table. In the latter case, we are explicitly
setting a new frame for this symbol table, e.g., a FunctionDeclaration
which will definitely setup a new stack frame.

The highwaterMark is replaced by simply checking the frame's storagePointer
before emitting the ENTER instruction.

Do note that this change means that we no longer reuse slots in the frame
for two variables with non-overlapping lifetimes. This can be reimplemented
later. The downside is increased runtime memory use. The upside is that
we're one step closer to not assigning storage locations until late in the
compile pipeline.