CompilerPassGenerics: Improve ergonomics of AST rewriting again
Rewriting a Block involves making a new instance which includes the change.
This can make it difficult to track a specific block as it is transformed
in the compiler. Solve this problem by adding an Id to Block which is
preserved through modifications to the AST.

Made changes to SymbolTable to associate each scope with a Block, by Id.
Clients may lookup the scope in which a symbol or type is defined. They
may lookup the Block Id associated with that scope. In this way, a compiler
pass such as CompilerPassGenerics may determine which block includes a
specific declaration.

Each expression which uses a generic type application must be rewritten to
instead refer to the concrete instantiation of the generic type. It may be
necessary to rewrite the AST to include a declaration for this concrete
instantiation. Now, the compiler will lean on the symbol table to tell it
the Id of the block in which the generic type is declared. It builds up a
list of pending rewrites to add to each block, by Id, and applies them as
it encounters each such block.