Workaround: must take special care when comparing StructType
I'm grappling with an issue where SymbolType.structType(StructType) can
persist in the AST across compiler passes, which leads to the type table
not containing an exact match.

Workaround this issue by taking special care when comparing the symbol
tables for two struct types: ignore declaration order and ignore
function symbols.

A better solution might be to either have SymbolType.structType refer
to the StructType through something similar to an identifier, (but it
cannot be the actual identifier) or by only referring to types in the
AST using type expressions and replacing PrimitiveType with a set of
type expressions for type atoms.