DeclScan: consumes a scope's scopePrologue and inserts code in the block
In order to setup vtables, SnapSubcompilerImplFor adds code to the
"scopePrologue" for the current scope. This may also happen, indirectly,
when the compiler invokes the type checker on a GenericTypeApplication
expression.

Now, when CompilerPassDeclScan compiles a block, it will take the
scopePrologue code from the block's scope, insert it into the block,
and then clear the scopePrologue, consuming it.

A couple of unit tests must be updated. The expected values have changed
due to different register numbering in the code sequence. This is due to
changes in the order in which code is compiled with this change.