Check validitity of trait conversion without checking for vtable instance
Previously, we would check the validity of a conversion from a trait to a
pointer-to-struct by checking that the struct type had the same name as the
corresponding trait vtable. However, there are cases where we need to check
the validity of the conversion before the vtable instance is compiled.
In order to handle these, the compiler will now check that the struct type
"conforms" to the required vtable type, which is to say that, for each
entry in the vtable, there is a corresponding member in the struct with an
appropriate type.