SnapBenchmark: Fix failing downcast to Seq
Some time ago, I changed SnapToTurtle16Compiler.tack to be a Result which
contains a TackProgram instead of a Seq. This introduced a build warning,
and runtime error, in SnapBenchmark when it tried to cast the result of
get() to Seq. Fix this by rewriting the caller to acknowledge TackProgram.