CompilerPassGenerics: instantiates generic functions during the pass
The Generics compiler pass instantiates generic functions itself. The
pass does this by partially evaluating the generic function with the
concrete type arguments supplied to the generic type application.

Since the language now utterly forbids shadowing, any identifier we find
in the current lexical scope must have exactly one definition in the
environment. We can iterate the function AST and replace all instances
of the identifier which refers to a generic type parameter with the
expression for the concrete type argument.