CompilerPassEraseVar is reworked into CompilerPassLowerVarDecl
Instead of erasing the VarDeclaration node, the compiler pass will extract
the expression so it can be treated specially. The VarDeclaration remains
in the AST so subsequent compiler passes can still scan the symbol.

CoreToTackCompiler is able to handle this and will erase Snap AST nodes
such as VarDeclaration that are entirely useless in Tack code.