Rework name mangling scheme to use a new "breadcrumb" trail
Each scope in the environment may have a breadcrumb which helps uniquely
identify function symbols within that scope. We ensure mangled names can
not collide with user-provided identifiers by using characters that the
parser does not permit for identifiers, such as '[' and ']' and ':'.

This replaces the existing system which used enclosingFunctionNameMode
and enclosingFunctionTypeMode to determine the enclosing scopes.