StructTypeInfo is updated in-place as the compiler traverses the AST
A struct gains methods when the compiler encounters Impl and ImplFor
nodes as it traverses the AST. These methods remain on the struct for
as long as the Impl and ImplFor nodes remain in scope. On exit from the
scope, these methods drop off the struct.

The compiler implements this behavior by adding a new symbol table to
StructTypeInfo. Methods are added to this table on encountering an Impl
or ImplFor node. A new mechanism is added to Env to record deferred
actions to be executed on exiting the scope. The compiler will arrange
to push a deferred action to remove the new methods.

Also, the vtable instance for the struct's conformance to a particular
trait is now inserted into the AST just after the ImplFor node.
Previously, it was inserted earlier in the program. However, the vtable
should also only be in scope where the methods are in scope. This is
necessary because it is possible to declare two independent conformances
to the same trait so long as this is done in entirely disjoint scopes.