CompilerPass: track expression evaluation context
The compiler evaluates expressions in various contexts, depending on where
the expression may be found in the AST. For example, in a concrete context,
an Identifier resolves to a sequence of instructions which loads the value
from memory; however, in a type context, the same expression is unchanged
by evaluation and refers to a type name.

The available contexts are concrete, temporary, type, and none.
concrete  -- Evaluating for a concrete value with the same type as the expr
temporary -- Evaluating for a pointer holding the address of the value.
type      -- Evaluating a type expression which resolves to a type
none      -- Expression is unmodified. Though, children may be evaluated in
             a different context, depending on which node we're examining.