CompilerPassClearSymbols: Change which avoids seeming compiler bug
For reasons I do not understand, the Swift compiler now sometimes
creates invalid, uninitialized objects when we use `as!` to down-cast.
This used to raise a runtime error. Now it leaves an invalid object that
causes a crash later.

As I run into these issues, I'm making changes at each site to try to
work around the strange issue.