Remove force-downcast from CompilerPass.visit(eseq:)
This is another case where it really does seem like the Swift compiler
is compiling force-downcast incorrectly. I find that a Seq may be erased
during "flattening" if it has one child, replaced by the child. In this
case, visit(seq:) may not return a Seq. My expectation is that a line
like `visit(seq: foo) as! Seq` would cause a runtime error when this
happens. However, this is not so. The compiler instead passes an invalid
Seq object which blows up the compiler with EXC_BAD_ACCESS sometime
later.

Changing the `as!` to `as?` avoids the problem, gives us more robust
error handling, and is just generally better.