MainBoard: Add series termination resistors and additional bulk capacitors
Clock and control signals generated on the main board now have series
termination resistors. These are specified now as zero ohm resistors, and
can be easily swapped for different values if needed.

There are pads for bulk capacitors next to each card slot. These are
specfied as Do Not Populate, and can be easily populated if needed.