eraseEseq: we cannot erase an Eseq used as a while-loop condition
The condition must be evaluated for side effects on each loop iteration,
which we cannot do if we erase the Eseq here. Instead, we'll have to
lower the While loop first in an earlier compiler pass.

Also, move eraseEseq pass closer to the end of compilation so it runs after
whilePass.