Type checker classes now support a `bypassConstAssignmentRules` options
The type checker classes now accept an options struct. The one option in
this today is `bypassConstAssignmentRules` which permits an assignment
to be accepted even when it violates rules prohibiting assignment to
initialized const variables.

This option is intended to be used as a temporary escape hatch to allow
CoreToTackCompiler to lower certain expressions in ways that generate
correct code, but which violate these rules.