Rename SnapSubcompilerVarDeclaration to VarDeclarationScanner
SnapSubcompilerVarDeclaration was only ever to process VarDeclaration nodes
for their effect on the environment, e.g., symbols, except for one place
where the generated assignment was actually used.

Rename the class to VarDeclarationScanner and change the interface so that
it only scans. The one place that did need an assignment now does that by
itself.